refactored App.js and added form component
