+++ src/App.js
@@ -1,20 +1,13 @@
 import React from "react";
-
-const todoList = [
-  { title: "This is a title", id: 1 },
-  { title: "Yet another title", id: 2 },
-  { title: "The last title", id: 3 },
-];
+import TodoList from "./todoList";
+import AddTodoForm from "./AddTodoForm";
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map((item) => (
-          <li key={item.id}>{item.title}</li>
-        ))}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
